Login/auth and autentication/crypting/same contacts on different accounts
Created a login route and logic for authentication and authorization of the following routes: list contacts, find contact by ID, and add new contact.
Implemented functionality that allows creation of a contact with the same name, email, or/and phone as long as the owner is unique, allowing users to have the same contacts if they know the same person.
Implemented password salting using bcrypt for increased security.
Made a few quality of life tweaks to improve the user experience.
Note: JWT was used for authentication, but Passport was not used in this implementation, although it may be used in the future.